+++ e2e/journeys/manageCases/caseProgression/checkApplicationSendToGateKeeper/sendToGateKeeper/sendToGateKeeperJourney.ts
