+++ e2e/tests/manageCases/caseProgression/application-journeys-check-gatekeeper-judge-order/application-journeys-check-gatekeeper-judgeUO-order.spec.ts
@@ -1,7 +1,7 @@
 import { test } from "@playwright/test";
 import Config from "../../../../config";
 import createDaCitizenCourtNavCase from "../../../../common/createCaseHelper";
-import { ApplicationJourneysCheckGatekeeperJudgeUOOrder } from "../../../../journeys/manageCases/caseProgression/application-journeys-check-gatekeeper-judge-DA-order/application-journeys-check-gatekeeper-judgeUO-order";
+import { ApplicationJourneysCheckGatekeeperJudgeUOOrder } from "../../../../journeys/manageCases/caseProgression/manageOrders/application-journeys-check-gatekeeper-judgeUO-order";
 import { Helpers } from "../../../../common/helpers";
 import config from "../../../../config";
